core/pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["srivatsan177 <ksrivatsan177@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mongoengine = "^0.27.0"
boto3 = "^1.34.21"
redis = "^5.0.1"
pyjwt = "^2.8.0"
fastapi = {extras = ["all"], version = "^0.109.0"}


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
